Refactor strategy class to integrate RobotClient functionality and enhance request handling. Removing RobotClient
